requirements.txt
ipdb==0.10.3
nose==1.3.7
jsmin==2.2.2
pyyaml==3.12
